hash

+++ libprg/src/libprg/hash.c
@@ -40,7 +40,7 @@ int hash(char* chave, int m) {
     return soma % m;
 }
 
-int inserir_hash(dicionario_t* d, char* chave, int valor) {
+int inserir_hash(dicionario_t* d, char* chave, long valor) {
     if (d == NULL || chave == NULL) return 1;
 
     int indice = hash(chave, d->tamanho);
@@ -72,7 +72,7 @@ int inserir_hash(dicionario_t* d, char* chave, int valor) {
     return 0;
 }
 
-int buscar_hash(dicionario_t* d, char* chave, int* valor) {
+int buscar_hash(dicionario_t* d, char* chave, long* valor) {
     if (d == NULL || chave == NULL || valor == NULL) return 1;
 
     int indice = hash(chave, d->tamanho);
